More Cleanup and Deduplication (#675)
