+++ backend/danswer/direct_qa/llm_utils.py
@@ -1,5 +1,3 @@
-from openai.error import AuthenticationError
-
 from danswer.configs.app_configs import QA_TIMEOUT
 from danswer.direct_qa.interfaces import QAModel
 from danswer.direct_qa.qa_block import QABlock
@@ -10,25 +8,6 @@ from danswer.llm.factory import get_default_llm
 from danswer.utils.logger import setup_logger
 
 logger = setup_logger()
-
-
-def check_model_api_key_is_valid(model_api_key: str) -> bool:
-    if not model_api_key:
-        return False
-
-    llm = get_default_llm(api_key=model_api_key, timeout=10)
-
-    # try for up to 2 timeouts (e.g. 10 seconds in total)
-    for _ in range(2):
-        try:
-            llm.invoke("Do not respond")
-            return True
-        except AuthenticationError:
-            return False
-        except Exception as e:
-            logger.warning(f"GenAI API key failed for the following reason: {e}")
-
-    return False
 
 
 # TODO introduce the prompt choice parameter
