+++ backend/danswer/server/search_backend.py
@@ -1,5 +1,3 @@
-from collections.abc import Generator
-
 from fastapi import APIRouter
 from fastapi import Depends
 from fastapi import HTTPException
@@ -9,30 +7,19 @@ from sqlalchemy.orm import Session
 
 from danswer.auth.users import current_admin_user
 from danswer.auth.users import current_user
-from danswer.configs.app_configs import DISABLE_GENERATIVE_AI
-from danswer.configs.app_configs import NUM_DOCUMENT_TOKENS_FED_TO_GENERATIVE_MODEL
-from danswer.configs.constants import IGNORE_FOR_QA
 from danswer.db.engine import get_session
 from danswer.db.feedback import create_doc_retrieval_feedback
-from danswer.db.feedback import create_query_event
 from danswer.db.feedback import update_query_event_feedback
 from danswer.db.models import User
 from danswer.direct_qa.answer_question import answer_qa_query
-from danswer.direct_qa.interfaces import DanswerAnswerPiece
-from danswer.direct_qa.interfaces import StreamingError
-from danswer.direct_qa.llm_utils import get_default_qa_model
-from danswer.direct_qa.qa_utils import get_usable_chunks
+from danswer.direct_qa.answer_question import answer_qa_query_stream
 from danswer.document_index.factory import get_default_document_index
 from danswer.document_index.vespa.index import VespaIndex
 from danswer.search.access_filters import build_access_filters_for_user
-from danswer.search.danswer_helper import query_intent
 from danswer.search.danswer_helper import recommend_search_flow
 from danswer.search.models import IndexFilters
-from danswer.search.models import QueryFlow
-from danswer.search.models import SearchQuery
 from danswer.search.search_runner import chunks_to_search_docs
 from danswer.search.search_runner import danswer_search
-from danswer.search.search_runner import search_chunks
 from danswer.secondary_llm_flows.extract_filters import extract_question_time_filters
 from danswer.secondary_llm_flows.query_validation import get_query_answerability
 from danswer.secondary_llm_flows.query_validation import stream_query_answerability
@@ -41,13 +28,10 @@ from danswer.server.models import QAFeedbackRequest
 from danswer.server.models import QAResponse
 from danswer.server.models import QueryValidationResponse
 from danswer.server.models import QuestionRequest
-from danswer.server.models import RerankedRetrievalDocs
 from danswer.server.models import SearchDoc
 from danswer.server.models import SearchFeedbackRequest
 from danswer.server.models import SearchResponse
-from danswer.server.utils import get_json_line
 from danswer.utils.logger import setup_logger
-from danswer.utils.timing import log_generator_function_time
 
 logger = setup_logger()
 
@@ -190,144 +174,10 @@ def stream_direct_qa(
     user: User | None = Depends(current_user),
     db_session: Session = Depends(get_session),
 ) -> StreamingResponse:
-    send_packet_debug_msg = "Sending Packet: {}"
-    top_documents_key = "top_documents"
-    unranked_top_docs_key = "unranked_top_documents"
-    predicted_flow_key = "predicted_flow"
-    predicted_search_key = "predicted_search"
-    query_event_id_key = "query_event_id"
-
-    logger.debug(
-        f"Received QA query ({question.search_type.value} search): {question.query}"
+    packets = answer_qa_query_stream(
+        question=question, user=user, db_session=db_session
     )
-    logger.debug(f"Query filters: {question.filters}")
-
-    @log_generator_function_time()
-    def stream_qa_portions(
-        disable_generative_answer: bool = DISABLE_GENERATIVE_AI,
-    ) -> Generator[str, None, None]:
-        answer_so_far: str = ""
-        query = question.query
-        offset_count = question.offset if question.offset is not None else 0
-
-        time_cutoff, favor_recent = extract_question_time_filters(question)
-        question.filters.time_cutoff = time_cutoff  # not used but just in case
-        filters = question.filters
-
-        user_acl_filters = build_access_filters_for_user(user, db_session)
-        final_filters = IndexFilters(
-            source_type=filters.source_type,
-            document_set=filters.document_set,
-            time_cutoff=time_cutoff,
-            access_control_list=user_acl_filters,
-        )
-
-        search_query = SearchQuery(
-            query=query,
-            search_type=question.search_type,
-            filters=final_filters,
-            favor_recent=favor_recent,
-        )
-
-        ranked_chunks, unranked_chunks = search_chunks(
-            query=search_query, document_index=get_default_document_index()
-        )
-
-        # TODO retire this
-        predicted_search, predicted_flow = query_intent(query)
-
-        if not ranked_chunks:
-            logger.debug("No Documents Found")
-            empty_docs_result = {
-                top_documents_key: None,
-                unranked_top_docs_key: None,
-                predicted_flow_key: predicted_flow,
-                predicted_search_key: predicted_search,
-            }
-            logger.debug(send_packet_debug_msg.format(empty_docs_result))
-            yield get_json_line(empty_docs_result)
-            return
-
-        top_docs = chunks_to_search_docs(ranked_chunks)
-        unranked_top_docs = chunks_to_search_docs(unranked_chunks)
-        initial_response = RerankedRetrievalDocs(
-            top_documents=top_docs,
-            unranked_top_documents=unranked_top_docs,
-            # if generative AI is disabled, set flow as search so frontend
-            # doesn't ask the user if they want to run QA over more documents
-            predicted_flow=QueryFlow.SEARCH
-            if disable_generative_answer
-            else predicted_flow,
-            predicted_search=predicted_search,
-            time_cutoff=time_cutoff,
-            favor_recent=favor_recent,
-        ).dict()
-
-        logger.debug(send_packet_debug_msg.format(initial_response))
-        yield get_json_line(initial_response)
-
-        if disable_generative_answer:
-            logger.debug("Skipping QA because generative AI is disabled")
-            return
-
-        try:
-            qa_model = get_default_qa_model()
-        except Exception as e:
-            logger.exception("Unable to get QA model")
-            error = StreamingError(error=str(e))
-            yield get_json_line(error.dict())
-            return
-
-        # remove chunks marked as not applicable for QA (e.g. Google Drive file
-        # types which can't be parsed). These chunks are useful to show in the
-        # search results, but not for QA.
-        filtered_ranked_chunks = [
-            chunk for chunk in ranked_chunks if not chunk.metadata.get(IGNORE_FOR_QA)
-        ]
-
-        # get all chunks that fit into the token limit
-        usable_chunks = get_usable_chunks(
-            chunks=filtered_ranked_chunks,
-            token_limit=NUM_DOCUMENT_TOKENS_FED_TO_GENERATIVE_MODEL,
-            offset=offset_count,
-        )
-        logger.debug(
-            f"Chunks fed to LLM: {[chunk.semantic_identifier for chunk in usable_chunks]}"
-        )
-
-        try:
-            for response_packet in qa_model.answer_question_stream(
-                query, usable_chunks
-            ):
-                if response_packet is None:
-                    continue
-                if (
-                    isinstance(response_packet, DanswerAnswerPiece)
-                    and response_packet.answer_piece
-                ):
-                    answer_so_far = answer_so_far + response_packet.answer_piece
-                logger.debug(f"Sending packet: {response_packet}")
-                yield get_json_line(response_packet.dict())
-        except Exception:
-            # exception is logged in the answer_question method, no need to re-log
-            logger.exception("Failed to run QA")
-            yield get_json_line(
-                {"error": "The LLM failed to produce a useable response"}
-            )
-
-        query_event_id = create_query_event(
-            query=query,
-            search_type=question.search_type,
-            llm_answer=answer_so_far,
-            retrieved_document_ids=[doc.document_id for doc in top_docs],
-            user_id=None if user is None else user.id,
-            db_session=db_session,
-        )
-
-        yield get_json_line({query_event_id_key: query_event_id})
-        return
-
-    return StreamingResponse(stream_qa_portions(), media_type="application/json")
+    return StreamingResponse(packets, media_type="application/json")
 
 
 @router.post("/query-feedback")
